+++ src/main/java/com/web/dao/WalletDao.java
@@ -3,6 +3,8 @@ package com.web.dao;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
+
 /**
  * 钱包dao
  * @author chxy
@@ -16,5 +18,5 @@ public interface WalletDao {
      * @param clickToken
      * @return
      */
-    boolean addToken(@Param("userId") int userId, @Param("clickToken") double clickToken);
+    boolean addToken(@Param("userId") int userId, @Param("clickToken") BigDecimal clickToken);
 }
